requirements/base.txt
Django==2.2
django-crispy-forms==1.8.1
djangorestframework==3.9
pytz
python-decouple
django-environ
africastalking
requests